[Dispatcher] Change zip folder structure for combined "full reports".
In OSX (and probably Windows) if a zip contains only one folder,
the name of the extracted result is not the name of the zip file,
but the name of the sole folder content. In this case the name
was e.g. "Regular" if a family had only one regular style. This
change results in the same folder if just the report for regular
was downloaded or if the combined report containing only one
entry for regular, e.g. "Diffentator_FamilyName_Regular".